Add identities
